+++ src/pages/dealershipDetails/components/CoApplicantsTable.js
@@ -7,8 +7,13 @@ import TableCell from '@material-ui/core/TableCell';
 import TableHead from '@material-ui/core/TableHead';
 import TableRow from '@material-ui/core/TableRow';
 import Typography from '@material-ui/core/Typography';
+import { useSnackbar } from 'notistack';
 import React, { useState } from 'react';
+import { useQueryClient } from 'react-query';
 import CreditInfoSideWrapper from './CreditInfoSideWrapper';
+import DeleteButton from '../../../components/CommonComponents/Button/DeleteButton';
+import { logger } from '../../../config/logger';
+import { deleteApplicantById } from '../../../services/dealers.service';
 
 const useStyles = makeStyles(theme => ({
   wrapper: {
@@ -38,7 +43,10 @@ const useStyles = makeStyles(theme => ({
     // paddingTop: 8
   },
   tableRow: {
-    cursor: 'pointer'
+    cursor: 'pointer',
+    '&:hover': {
+      backgroundColor: '#fafafa',
+    }
   },
   document: {
     display: 'inline-block',
@@ -52,7 +60,35 @@ const useStyles = makeStyles(theme => ({
 
 const CoApplicantsTable = ({id, editable, coApplicantsData, titleAlign, getExperianData, onClickAddMenu, formType, openCloseCreditForm, currentUser, showDealerEditForm, dealersClickRow, editFormClose }) => {
   const classes = useStyles();
-  const [rowData, setRowData] = useState()
+  const queryClient = useQueryClient()
+  const { enqueueSnackbar } = useSnackbar();
+  const [rowData, setRowData] = useState();
+  const [deleteModal, setDeleteModal] = useState(false);
+
+  const DeleteApplicant = (row_data) => {
+    deleteApplicantById(id, row_data?.id, row_data?.userType)
+      .then(res => {
+        queryClient.invalidateQueries(['co-applicants', id])
+        setDeleteModal(false)
+        enqueueSnackbar(res.message, {
+          anchorOrigin: {
+            vertical: 'top',
+            horizontal: 'right',
+          },
+          variant: 'success',
+        })
+      })
+      .catch(e => {
+        logger(e)
+        enqueueSnackbar(e, {
+          anchorOrigin: {
+            vertical: 'top',
+            horizontal: 'right',
+          },
+          variant: 'error',
+        })
+      })
+  }
     
   if (!coApplicantsData || !coApplicantsData.length)
     return (
@@ -86,7 +122,7 @@ const CoApplicantsTable = ({id, editable, coApplicantsData, titleAlign, getExper
           </TableRow>
         </TableHead>
         <TableBody>
-          {coApplicantsData.map(row => (
+          {coApplicantsData.map((row, index) => (
             <TableRow className={classes.tableRow} key={row.id}>
               <TableCell onClick={e => editable && dealersClickRow(e, row, 'COAPPLICANT')}>
                 {row.first_name}&nbsp;&nbsp;
@@ -116,6 +152,7 @@ const CoApplicantsTable = ({id, editable, coApplicantsData, titleAlign, getExper
                 editable &&
                   <TableCell align="right">
                     <Button size='small' variant='outlined' color='secondary' onClick={() => setRowData(row)}>Credit Info</Button>
+                    <DeleteButton alertText={`Do you really want to delete this co-applicant named ${row?.first_name}?`} deleteAction={() => DeleteApplicant(row)} deleteModal={deleteModal} setDeleteModal={setDeleteModal} id={index} buttonType='icon' />
                   </TableCell>
               }
             </TableRow>