+++ src/pages/dealershipDetails/components/GuarantorsTable.js
@@ -8,8 +8,13 @@ import TableCell from '@material-ui/core/TableCell';
 import TableHead from '@material-ui/core/TableHead';
 import TableRow from '@material-ui/core/TableRow';
 import Typography from '@material-ui/core/Typography';
+import { useSnackbar } from 'notistack';
 import React, { useState } from 'react';
+import { useQueryClient } from 'react-query';
 import CreditInfoSideWrapper from './CreditInfoSideWrapper';
+import DeleteButton from '../../../components/CommonComponents/Button/DeleteButton';
+import { logger } from '../../../config/logger';
+import { deleteApplicantById } from '../../../services/dealers.service';
 
 const useStyles = makeStyles(theme => ({
   wrapper: {
@@ -28,7 +33,10 @@ const useStyles = makeStyles(theme => ({
     marginBottom: 8
   },
   tableRow: {
-    cursor: 'pointer'
+    cursor: 'pointer',
+    '&:hover': {
+      backgroundColor: '#fafafa',
+    }
   },
   sidePanelWrapper: {
     width: '40vw',
@@ -38,7 +46,35 @@ const useStyles = makeStyles(theme => ({
 
 const GuarantorsTable = ({ id, editable, guarantorsData, titleAlign, getExperianData, onClickAddMenu, formType, openCloseCreditForm, currentUser, showDealerEditForm, dealersClickRow, editFormClose }) => {
   const classes = useStyles();
-  const [rowData, setRowData] = useState()
+  const queryClient = useQueryClient()
+  const { enqueueSnackbar } = useSnackbar();
+  const [rowData, setRowData] = useState();
+  const [deleteModal, setDeleteModal] = useState(false);
+  
+  const DeleteApplicant = (row_data) => {
+    deleteApplicantById(id, row_data?.id, row_data?.userType)
+      .then(res => {
+        queryClient.invalidateQueries(['guarantors', id])
+        setDeleteModal(false)
+        enqueueSnackbar(res.message, {
+          anchorOrigin: {
+            vertical: 'top',
+            horizontal: 'right',
+          },
+          variant: 'success',
+        })
+      })
+      .catch(e => {
+        enqueueSnackbar(e, {
+          anchorOrigin: {
+            vertical: 'top',
+            horizontal: 'right',
+          },
+          variant: 'error',
+        })
+        logger(e)
+      })
+  }
 
   if (!guarantorsData || !guarantorsData.length)
     return (
@@ -71,7 +107,7 @@ const GuarantorsTable = ({ id, editable, guarantorsData, titleAlign, getExperian
           </TableRow>
         </TableHead>
         <TableBody>
-          {guarantorsData.map(row => (
+          {guarantorsData.map((row, index) => (
             <TableRow className={classes.tableRow} key={row.id}>
               <TableCell onClick={e => editable && dealersClickRow(e, row, 'GUARANTOR')}>
                 {row.first_name}&nbsp;&nbsp;
@@ -82,6 +118,7 @@ const GuarantorsTable = ({ id, editable, guarantorsData, titleAlign, getExperian
                 editable &&
                   <TableCell align="right">
                     <Button size='small' variant='outlined' color='secondary' onClick={() => setRowData(row)}>Credit Info</Button>
+                    <DeleteButton alertText={`Do you really want to delete this guarantor named ${row?.first_name}?`} deleteAction={() => DeleteApplicant(row)} deleteModal={deleteModal} setDeleteModal={setDeleteModal} id={index} buttonType='icon' />
                   </TableCell>
               }
             </TableRow>