+++ CHANGELOG.md
@@ -2,6 +2,53 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.10.0](https://gitlab.com/vignesh3/petromall-fe/compare/v2.9.0...v2.10.0) (2022-02-22)
+
+
+### Features
+
+* **delete-applicant:** option to delete applicants in dealership ([060e419](https://gitlab.com/vignesh3/petromall-fe/commit/060e4197e7ea0e89238ebf6108c93b89032d29d6))
+
+
+### Bug Fixes
+
+* **pending-approval:** submitted form in pending approval ([973c2a7](https://gitlab.com/vignesh3/petromall-fe/commit/973c2a7e4e453b22cff5099e510d21114a835887))
+* **sign-form:** sign form close issue ([a78a9c7](https://gitlab.com/vignesh3/petromall-fe/commit/a78a9c7cbc8fff5953c7c33dd1ad697097911a47))
+
+## [2.9.0](https://gitlab.com/vignesh3/petromall-fe/compare/v2.8.3...v2.9.0) (2022-02-16)
+
+
+### Features
+
+* **credit-withheld:** highlight withheld loans ([2106355](https://gitlab.com/vignesh3/petromall-fe/commit/210635510cf8c8b7117c07b828f7e3a2c5384ae3))
+
+
+### Bug Fixes
+
+* **dealer-passbook:** invalid date format ([dfb9668](https://gitlab.com/vignesh3/petromall-fe/commit/dfb9668d34cd70016b9ab8877fda5efc9439f750))
+* **disb-details:** date format and prefilled data ([0de4625](https://gitlab.com/vignesh3/petromall-fe/commit/0de4625c5941161a318fc2048d76517184301d72))
+* **name-regex:** allow spaces in name ([00f86c2](https://gitlab.com/vignesh3/petromall-fe/commit/00f86c2a918957dc215ab25f49cb026344991853))
+* **styles:** update table style ([87c863b](https://gitlab.com/vignesh3/petromall-fe/commit/87c863bf89767f947265cfc828585a38c208ed43))
+
+### [2.8.3](https://gitlab.com/vignesh3/petromall-fe/compare/v2.8.2...v2.8.3) (2022-01-28)
+
+
+### Bug Fixes
+
+* merge msg ([5e26082](https://gitlab.com/vignesh3/petromall-fe/commit/5e26082241794f0de70eedd898eb2eb0ba5023f3))
+* Merge Msg ([4257dc9](https://gitlab.com/vignesh3/petromall-fe/commit/4257dc9acea82442a614749205dfaf1a060782ff))
+
+### [2.8.2](https://gitlab.com/vignesh3/petromall-fe/compare/v2.8.1...v2.8.2) (2022-01-26)
+
+
+### Bug Fixes
+
+* **credit-reload:** remove unused code ([399304a](https://gitlab.com/vignesh3/petromall-fe/commit/399304ad2cc7f56460aede014bc9061fc6c5b3c2))
+* Merge Message ([678ab4b](https://gitlab.com/vignesh3/petromall-fe/commit/678ab4b5c83056233129677af146dca19060cc46))
+* merge msg ([2fcfea5](https://gitlab.com/vignesh3/petromall-fe/commit/2fcfea582ed37b9818f19ae9f8213d4763faf654))
+* merge msg ([2c393d3](https://gitlab.com/vignesh3/petromall-fe/commit/2c393d352846a24711dae83c3ee73a55d371e154))
+* **pd:** add rule to credit head ([155beeb](https://gitlab.com/vignesh3/petromall-fe/commit/155beeb9fd71a6ce99aafb697afa10b495bad3f5))
+
 ### [2.8.1](https://gitlab.com/vignesh3/petromall-fe/compare/v2.8.0...v2.8.1) (2022-01-17)